Create Flip string to monotone increasing (Leetcode)
Intuition
Approach 1:
While traversing over the string, we'll count the number of ones and also count no. of zeroes.
The only difference is.. if we will append a one in a string that is already a monotone.. then it will not create any difference.
But if we append zero .. then we need to flip all ones before it.
So we'll update in case of zero counter as... countzero =min(countzero,countone).
Approach 2:
We can solve it by using...
Left and right window.
Initially, the size of lw(left window) is zero.. and lw will carry the number of ones present in the window

Similarly
Rw will contain the no. Of zeros.. initially it is of size =n

.. and a variable sum =0 initially..

Sum = min(lw+rw , sum)..

And in every pass, we'll increase the size of lw by one.. and decrease the size of rw by one... And update the sum...

At last, we'll return the sum.

Complexity
Time complexity: O(1)
Space complexity: O(1)