Create Clone_Graph(leetcode)
  This is the best explanation i found on leetcode...
it's his solution.
						   
OBSERVATIONS
1. Cloned Graph has all nodes as newnodes and doesn't have node as a reference to given original graph.
2. Graph has no practical edges, i.e no pointers for edges.
         (2,3)            (1,3)<---adj list
		  (1)-------------(2)
		    |              |
			|              |
			|              |
		  (4)-------------(3)
		  (1,3)          (2,4)
If I say '1' is my starting point and how should I jump to '2' for that I have to iterate through this adjacency list.
ALGORITHM TO USE
We need to traverse all node of original graph and as soon as we reach a node, we will make a copy node.
And recur for the rest of the graph.
This is a typical recursion-type problem implemented on Graph.
For 'Recursion' we use basically 'DFS' or 'BFS'.
I am using DFS
KEY POINTS
1. We use HashMap to solve it and use DFS.
2. Initially our hash map will be empty and we try to map the old node with the new node or the copy node.
3. We start with any entry point.
4. I am using '1' as my entry point.

Now let's see how it's going I am starting with 1 and whenever I visited a new node I cloned it and put it in there.
We are using DFS so the algorithm is like 'it starts at the root node (select some arbitrary node as the root node in the case of a graph) and explores as far as possible along each branch before backtracking.
So the basic idea is to start from the root or any arbitrary node and mark the node and move to the adjacent unmarked node and continue this loop until there is no unmarked adjacent node. 
Then backtrack and check for other unmarked nodes and traverse them. Finally, print the nodes in the path.'
So we are using HashMap to put all the visited nodes or old nodes there with clone one too.
 _________
| HashMap |
 ----------
|Old|Clone| 
| 1  | 1 |
|  2 | 2 |
|  3 | 3 |                                                     
|  4 | 4 |                                                                                                            
---------

Now I started with 1 so I cloned it and from 1 can go to 2 and 4 so I go 2 and when I visited 2 I cloned 2 and now I have two choices either go to the previous one is 1 or discover 3 i.e new node
so accordingly to dfs i go to 3 and from 3 I can go to 4 I go there and cloned it. Now if we see for each node we have visited to a new node but what about 4? So here half part of Dfs is completed,
and now it's time for the recursive call to go back, and now from here we check from the current node I can go where and where.
And follow the same rules over there.

BUT BEFORE STARTING ANY CLONING WE HAVE TO CHECK THAT IF WE HAVE CLONED THAT NODE ALREADY THERE OR NOT. IF NOT THAN ONLY WE CLONED IT. 

That's the only reason we are using a hash map so that we don't need to clone again and again.
For every uncloned node we make a clone and iterate over the neighbors of an original node using dfs traversal or bfs traversal.