Create Subarrays_Div_By_k(leetcode)_19-jan
HASH_MAP:

The solution makes use of a hashMap and works in a single pass.
The solution section gives something that is confusing for me so here is a more understandable solution:

curr_sum here is denoting the traditional cumulative sum at any point during the traversal of the arrray
Now curr_sum%k at any point gives the remainder,this can either be 0 or any other number.
Whatever the remainder is we are simply putting in the map thereby preparing a mapping of all the possible remainders from all the possible sum and storing their frequency as the values.
The only thing and the most important this is to understand why this line of code works and why it gives the desired outputs.
if(mp.find(rem)!=mp.end())
count+=mp[rem];
We are increasing the count by the number of times the given remainder has been seen.
I am considering it a fact that :
If the remainders of array upto i and array upto j are equal (ri=rj),
then subarray from i to j has a remainder of 0,
which means It's the type of subarrays we're looking for!
Look here: https://leetcode.com/problems/subarray-sums-divisible-by-k/discuss/413234/DETAILED-WHITEBOARD!-BEATS-100-(Do-you-really-want-to-understand-It)
Now lets dry run the code with this concept to drill it in.

A = [4,5,0,-2,-3,1], K = 5
A = 4, curr_sum = 4, curr_sum%K = 4 map now is: 0-->1, 4-->1 RESULT: []

A = 5, curr_sum = 9, curr_sum%K = 4 map now is: 0-->1, 4-->2 So this means we found a remainder that was already there! count+=1 RESULT: ([5], ) (added one more)

A = 0, curr_sum = 9, curr_sum%K = 4 map now is: 0-->1, 4-->3 So this means we found a remainder that was already there! count+=2 RESULT: ([5,0], [5], [0]) (added two more)

A = -2, curr_sum = 7, curr_sum%K = 2 map now is: 0-->1, 4-->3, 2-->1
RESULT: ([5,0], [5], [0])

A = -3, curr_sum = 4, curr_sum%K = 4 map now is: 0-->1, 4-->4, 2-->1 So this means we found a remainder that was already there! count+=3 RESULT: ([5,0], [5], [0], [5,0,-2,-3], [0,-2,-3], [-2,-3]) (added three more)

A = 1, curr_sum = 5, curr_sum%K = 0 map now is: 0-->2, 4-->4, 2-->1 So this means we found a remainder that was already there! count+=1 RESULT: ([5,0], [5], [0], [5,0,-2,-3], [0,-2,-3], [-2,-3], [4,5,0,-2,-3,1]) (added one more)

// VECTOR_APPROACH:

Understanding the Question
Subarray Sums divisible by k. Let's break the question down in small steps.

Firstly, we need subarray sums. There are n*(n-1)/2 of these for an n-length array.
Secondly, we need to count the number of subarray sums divisible by k (given).
So the first approach that inevitably comes to mind is to loop over all the possible subarrays, taking their sums and checking if they are divisible by k. But as soon as this comes to one's mind, one must also realize that this is an extremely naive solution as it is.
We would need to choose the starting and ending points (O(n) each) and then loop over the subarray to calculate its sum (again O(n)), giving a total of O(n3)O(n^3)O(n 
3
 ) time complexity, which is not permitted by the constraints given.

The next thing one may think is of using prefix sum array, to reduce the calculation of the sum to O(n2)O(n^2)O(n 
2
 ), but this will still not make the cut.

Look for Hints in the Constraints
The constraints are as follows:
1<=n<=3∗1041 <= n <= 3*10^41<=n<=3∗10 
4
 
−104<=nums[i]<=104-10^4 <= nums[i] <= 10^4−10 
4
 <=nums[i]<=10 
4
 
2<=k<=1042 <= k <= 10^42<=k<=10 
4
 

As we can see, the constraints won't permit an O(n2)O(n^2)O(n 
2
 ), and hence we need to make use of all the information in the question in such a way that we can do this in O(nlogn)O(nlogn)O(nlogn) or less.

A Curious Observation
Let us consider the following example:
nums = [4,5,0,-2,-3,1], k = 5
If we consider the prefix sum of the above array and the remainders when they are divided by k (i.e prefixSum[i]%k), we get the following:

prefixSum = [4, 9, 9, 7, 4, 5]
   ..%k   = [4, 4, 4, 2, 4, 0]
For any subarray [i..j] whose sum is divisible by k, we can easily see that prefixSum[j] - prefixSum[i-1] = 0.
Thus if there are any repetitions in our prefixSum%k array, that represents a required subarray.

So if we have the count of every possible value of prefixSum%k, we can determine if there are any subarray that end at the current index and the sum is divisible by k.