Create FindTownJudge||Celebrity_problem(leetcode)
This is the same as Find the Celebrity (Medium), but here we can access all connections directly (so this problem is Easy).

Brute-Force Solution
Since we have all connections, we can count people that i-th persontrusts, and people that i-th person is trusted by. Then, we just search for a judge.

int findJudge(int N, vector<vector<int>>& trust) {
  vector<int> trusts(N + 1), trusted(N + 1);
  for (auto &t : trust) ++trusts[t[0]], ++trusted[t[1]];
  for (auto i = 1; i <= N; ++i) if (trusts[i] == 0 && trusted[i] == N - 1) return i;
  return -1;
}
We can further simplify this by using a single vector and tracking the trust balance. Only the person that has N - 1 trust balance is trusted by everyone and trusts none.

int findJudge(int N, vector<vector<int>>& trust) {
  vector<int> balance(N + 1);
  for (auto &t : trust) --balance[t[0]], ++balance[t[1]];
  for (auto i = 1; i <= N; ++i) if (balance[i] == N - 1) return i;
  return -1;
}
Complexity Analysis
Time: O(m), where m is the number of trust pairs.
Memory: O(N).

Find the Celebrity
If we are given trust connections as an adjacency matrix (or a hash map), we can use the same algorithm as in the Find the Celebrity problem. Here is solution and explanations to that problem. This cool technique to quickly find a potential celebrity helps reduce the runtime and memory complexity.

int findJudge(int N, vector<vector<int>>& trust) {
  vector<vector<int>> knows(N + 1, vector<int>(N + 1));
  for (auto &t : trust) knows[t[0]][t[1]] = 1;
  return findCelebrity(N, knows);
}
int findCelebrity(int n, vector<vector<int>>& knows, int i = 1) {
  for (auto j = i + 1; j <= n; ++j) if (knows[i][j]) i = j;
  for (auto j = 1; j < i; ++j) if (knows[i][j]) return -1;
  for (auto j = 1; j <= n; ++j) if (i != j && !knows[j][i]) return -1;
  return i;
}
Complexity Analysis
Note that we analyze the complexity of findCelebrity (without preparation steps).

Time: O(N).
Memory: O(1).

69
69

Your description: This is the same as Find the Celebrity (Medium), but here we can access all connections directly (so this problem is Easy) is not right. Actually for both problems, we are given all connections. If we are not given all connections, how can we solve this question?
The real difference is that:

For Find Celebrity, we are given "adjacency matrix" boolean[n][n], we can get the value by asking knows(a, b). We can know if anyone knows anyone else or not.
For this problem, we are given edges only. From these edges, we only know who knows who, we cannot directly know if someone does not know another, it is implicit. For example: [1, 2], [2, 3], we know 1 knows 2 and 2 knows 3. But we cannot directly know if 1 knows 3 unless we make sure there is not an edge [1, 3] existed which requires some extra operations(like build graph by creating adjacency matrix new boolean[n][n] or HashSet to check). So it is not easy to solve this problems using Find Celebrity algorithm.



Thanks. I wasn't aware that you could do more than one operation in the body of a for loop without indenting it with braces. I generally prefer not to apply braces on loops containing a single(double) statement, as just the indendtation is more aesthetic than braces in this case.
int findJudge(int N, vector<vector<int>>& trust) {
  vector<int> trusts(N + 1), trusted(N + 1);
  for (auto &t : trust) ++trusts[t[0]], ++trusted[t[1]];
  for (auto i = 1; i <= N; ++i) if (trusts[i] == 0 && trusted[i] == N - 1) return i;
  return -1;
}