Create Maximum_Number_of_Toys(GOOGLE)(GFG).cpp
Approach to solve this problem:

1. As we need to find maximum number of toys that we can buy with a given amount. It's always preferable to go for toys with lowest cost.

Hence, we sort the array and start buying from index 1 to N till the money left is more than the cost of current toy.
This step takes O(N) time. We can do it in O(logN) using prefix and binary search.

2. As each query gives the indices of toys that are broken, and sorting will disturb the original indices, we need to store the original index with the cost.
Hence, we create an array of pair<int,int> to store cost and index.
Then we sort the new array.

3. Now, when we purchase from index 0 to N-1 of sorted array untill the money is left, there are two cases, either we encounter the broken toys indices or we don't.
We need a quick way to check where in the sorted array lies the broken toy.
Hence, we create a map<int,int> from original index to sorted index.
If broken toy's index in sorted array is after the last toy we purchase, we need not do anything.
But, if broken toy's index in sorted array is equal to of less than the last index of purchase. Then we remove that toy from our purchase and add the cost in our remaining money.
Now, the remaining money can we used to purchase other toys. And, as we have already gone through all the indices till last purchase, we only have check if we can buy the toys after that.
Condition to check, cost<= remaining money and index!=broken toy's index.

That's it, we store the solution for current query in a vector and return the vector.