Create LFU_Cache(Leetcode).cpp
Intuition of this Problem:
for, get(key) -->
gets the value of the key if exists, else -1;

for, put(key, value) -->
condition 1: updates the value of the key if present, else inserts the key directly.
condition 2: when cache is full -->
case i : if cache is full, removes the LFU guy
case ii : if cache is full, but exists multiple LFU guys (or a tie exists), then remove the LRU one and then add new node.

NOTE - PLEASE READ APPROACH FIRST THEN SEE THE CODE. YOU WILL DEFINITELY UNDERSTAND THE CODE LINE BY LINE AFTER SEEING THE APPROACH.

Approach for this Problem:
The class LFUCache has three unordered_map data structures:
keyVal: stores the key-value pairs, with the key as the first element of the pair and the value as the second element.
freqList: stores the keys in a list, with the key's frequency as the index of the list.
pos: stores the position of each key in the freqList.
The class also has a private integer variable 'capacity' that is initialized in the constructor and represents the maximum number of key-value pairs that can be stored in the cache.
The get() function is used to retrieve a value from the cache given a key. If the key is not found in the keyVal map, -1 is returned. Otherwise, the following steps are taken:
The key's current frequency is retrieved from keyVal.
The key is removed from the list at index equal to its current frequency in the freqList map.
The key's frequency in keyVal is incremented by 1.
The key is added to the list at index equal to its new frequency in the freqList map.
The key's position in the freqList map is updated in the pos map.
If the list at index equal to the current minimum frequency is empty, the minimum frequency is incremented by 1.
The value associated with the key is returned.
The put() function is used to add a new key-value pair to the cache or update the value of an existing key.
If the capacity is 0, the function returns immediately.
If the key already exists in the keyVal map, the following steps are taken:
The key's value is updated in the keyVal map.
The key is removed from the list at index equal to its current frequency in the freqList map.
The key's frequency in keyVal is incremented by 1.
The key is added to the list at index equal to its new frequency in the freqList map.
The key's position in the freqList map is updated in the pos map.
If the list at index equal to the current minimum frequency is empty, the minimum frequency is incremented by 1.
If the key does not already exist in the keyVal map and the size of keyVal is equal to the capacity, the following steps are taken:
The key at the front of the list at index equal to the current minimum frequency in the freqList map is removed from all three maps (keyVal, freqList, pos).
A new key-value pair is added to keyVal, with a frequency of 1.
The key is added to the list at index 1 in the freqList map.The key's position in the freqList map is updated in the pos map.
The minimum frequency is set to 1.