Create Best team with no conflict(Leetcode).cpp
Anytime I couldn't solve a problem, the above question bugged me alot. I have a train of thought which I use to get to the solution. Feel free to pick what works for you and device your own methods to arrive at solutions.

Thinking forward
In this train of thought, we first look at all information that is available to us. We are given 2 arrays of scores and age and the problem suggests that it has something to do with ordering w.r.t ages.
The word ordering quickly suggests that we can sort the input based on ages. ( But what kind of sort ? Ascending or Descending ? )
Will this sorting even take us closer to the solution ?

Jot down an example to answer yourself.
Let the scores array and ages array be the following ( Try to be as random as possible ) :
scores : [ 10, 2, 6, 4, 9 ]
ages. : [ 5, 1, 4, 5, 7]

Sorting the above array it by age (ASCENDING) leads me to ( If ages are equal sort ascending by score )
scores : [ 2, 6, 4, 10, 9 ]
ages. : [ 1, 4, 5, 5, 7]

Now that we get here, let's observe the constraints again.
Team is not allowed to have conflicts. A conflict exists if a younger player has a strictly higher score than an older player. A conflict does not occur between players of the same age.

Let's modify this statement for a pair of sorted arrays. Since the arrays are now sorted by age, any younger player comes before an older player. This means, IF I am choosing a bunch of players for my solution, THEN, in the order from left to right in scores array, I am allowed to choose scores as long as they are increasing --> More strictly, they arenon-decreasing because it is allowed for a score of younger player = score of older player.
With the scores being [ 2, 6, 4, 10, 9 ]
I can pick a team with [2, 4, 9] => Total = 15
I can pick a team with [2, 6, 10] => Total = 18
Note: In doing the above, I DID NOT consider the age array at all! Because, inherently, age of any element on the left is less than or equal to the ones of the right. (For i < j, age[i] <= age[j] )

Now I can reformulate the problem statement saying,
We are given one array (HERE Scores array) and we can choose any number of elements in the order from LEFT to RIGHT , PROVIDED that the sequence of elements I pick are non-decreasing.

To put it more formally,
We have to select a sequence S in the given array A, such that for this sequence S any pair of indices i, j should yeild S[i] <= S[j], and we have to maximize sum(S)

So far we have simplified whatever information that was given to us. Let's now go to the next step.

Thinking backwards
In this train of thought, we think What topics is this problem related to ?
Brainstorm your ideas in this step to find what is relevant to this problem. More formally, the questions you ask yourself are ( the following are random ideas just to see what sticks )
Can I apply two pointers ? ( Doesn't look like it! )
Can I think of graphs ? ( To be honest, I thought about this to be graph problem for a long time. I thought is it about finding longest path where edges are given between younger player and older player based on age and score constraints. But this didn't lead me anywhere )
I see sequence in my problem statement, does this involve any Longest Increasing Subsequence Methods ? (But again LIS is for increasing sequence, but I need non-decreasing, Do I need to tweak LIS probably? ( You just hit Jackpot! Congratulations!! )

Here, if you see my above train of thoughts, in thinking about graphs I was going orthogonal to the solution. IF you see yourself doing this, it is totally fine. What is important is that after you have solved the problem, DON'T jump to the next problem yet. TRY to fix your train of thoughts. Ask yourself, where could you improve? What questions should I ask myself so that I push myself closer to the solution ? What conclusions did I reach that I dropped the idea that Graphs isn't the right solution ? How can I reach this conclusion faster ?

At first, the answers to above questions are difficult to reach, but with practice, this will become second nature and your direction of thoughts will deviate less from the intended solution.

Constraints
BEFORE, we jump to the solution, one IMPORTANT aspect of problems are the constraints. Experienced programmers always look at constraints first.
In this case, let us see what information can be derived from the constraints given to us.
We have ,

* 1 <= scores.length, ages.length <= 1000 
* scores.length == ages.length
* 1 <= scores[i] <= 106
* 1 <= ages[i] <= 1000
These constraints give us answers to the following questions:

What solutions are admissible ?
The length of the input arrays is atmost 1000, ages.length and scores.length <= 1000 MAX.
Considering that approx 10^8 computations is 1 second of processing. We can see that O(n^2) solution is admissible because 1000*1000 = 10^6 which is approx 0.01 seconds.
However, had scores.length been 10^5, then O(n^2) solution is not admissible as n^2 = (10^5)^2 = 10^10 ~ 100 seconds. And we need an O(n log n) or O ( n root n) solution instead.
What data type should I be using to store my answer ?
TO answer this, let's see what is the maximum value of answer we can get, and compare that with max value different data types can hold.
scores[i] <= 10^6, and scores.length <= 1000 this means that the max answer we can get is 10^6 * 1000 = 10^9. ( A test case where all values are equal to 10^6 and all age are equal, allowing us to select the whole bunch as the answer ). An INT type can store this max value easily within itself.
However, IF scores[i] <= 10^7 , then in this case the max value for our answer becomes 10^10, suggesting that using INT will cause an overflow, and we need to use LONG LONG type to manage overflows.
A good rule of thumb, is to look at constraints first to arrive at a reasonable time complexity, then think forward to simplify the problem, then think backward to see what kind of problem I am looking at.